Remove useless and dangerous shutdown handler
As the shutdown handler contains a reference to the TCPDF instance itself, it prevents the destructor from being called.
In turn, the restoration of the mb_internal_encoding does not work correctly.

This affects all code that is executed after interaction with TCPDF. Code which uses multi-byte functions is using the wrong encoding (ASCII instead of the configured one). This might lead to wrong string operation results and as a consequence to data corruption.

I found the issue in an application which generates PDF reports. The mutation of global state (mb_internal_encoding) due to the bug led to test failures in another part of the application which is using multi-byte functions.

The destructor also calls the _destroy method, thus cleanup still works after removal of the shutdown handler.
A bonus of the removal is reduced memory usage, as the memory for the TCPDF instance is freed as soon as the object is not referenced anymore.
For an explicit cleanup (and in turn mb_internal_encoding restoration), developers may call unset($pdf) before other code execution.

PR #90 contains a minimal example to reproduce the faulty and fixed behavior.