Fix XRef parsing issue
When parsing XRef stream it assumes column length and a predictor is
set, if they’re not columns is set to 0 and the wrong chunk length is
read.

This change works out the correct column length (the sum of the /W
widths), and prepends a 0 column to every row to ensure the code below
it works correctly.