Change/fix named destination implementation
We noticed that named destination links created by TCPDF worked when the resulting PDFs were viewed in Google Chrome's built-in PDF viewer but not in Acrobat.  (In Acrobat, clicking on the link did nothing.  If I right-click on a link and choose Edit, Acrobat says that the action is to "Go to a page in this document," but it doesn't give any details or let me edit it - as if it doesn't know how to interpret the link's contents.)

From comparison with links created by Acrobat itself, and from comparison with unofficial named destination support for FPDF (http://www.fpdf.org/en/script/script99.php), this update seems to fix the problem.