URI in PDF can result in E_NOTICE
Generating a PDF file with a hyperlink like "tel:12345" results in an error: "Errstr: Undefined index: path" due to the fact that parse_url should only be used for URL not URI. 

SOURCE: https://www.php.net/manual/en/function.parse-url.php
"This function is intended specifically for the purpose of parsing URLs and not URIs."

This fix does not prevent parsing the link but checks if the parsed result contains a 'path'. If it does not than the rest of the if statement is ignored and no error is generated.