Problem with imagekeys undefined or unlinked
PR #136 removed problem with undefined `$imagekeys`, but the solution skipped cleaning files in `$imagekeys`.
The problem about this is that `_destroy` function is called 3 times (from `Close` function, as `shutdown_handler` and from `__desctruct` function). Only first time is called as `destroyall == false`, but unset `$imagekeys`. My solution is to preserve `$imagekeys` to be cleaned by call with `$imagekeys == true`.