GetCssBorderStyle Has Problem When !important Is Specified (#467)
* GetCssBorderStyle Has Problem When !important Is Specified

This is, in particular, a problem for PhpSpreadsheet. See https://github.com/PHPOffice/PhpSpreadsheet/issues/1164. TCPDF thinks that there can be 0, 1, 2, or 3 components to shorthand border style, but there can be more, e.g. !important. Logic is rearranged to allow for more than 3. A new example file is added to demonstrate that the fix works correctly.

* Improved Example

Better code coverage.

Co-authored-by: Nicola Asuni <nicolaasuni@users.noreply.github.com>