Extract method to get TCPDF producer
Wraps TCPDF_STATIC::getTCPDFProducer() calls with a private method and
adds an empty private property that can be used to change the producer
line. The private property can be changed with the reflection API.

Both method and property are declared as private to avoid BC breaks.

The idea is to help with testing PDF generation, by using a predictable
producer line.

Signed-off-by: Maurício Meneghini Fauth <mauricio@fauth.dev>