Fixed problems with S25 barcode
Using your S25 barcode generator I found a problem. The representation of the digit 8 was the same as 0. I fixed it. Now it is recognized correctly.

Additionally you were using different bar widths for the start and stop values. I changed this too.

One thing I did not change till now is the part where you add an leading 0 if the string is odd. Why do you do so?