fix: base64 images not rendered inside of html
I'm not sure when this happened, but default base64 encoded images were no longer rendered in html  text. 

If the image is rendered like this:
```
<img src="data:image/png;base64,....." />
```

the image was not rendered when inside of an html text. Instead one has to replace `data:image/png;base64,` with `@`which in my opinion is not ideal and an overhead. 

https://github.com/tecnickcom/TCPDF/issues/179